Move the Display model tests into the common lib
